+++ internal/config/reader.go
@@ -2,29 +2,74 @@ package config
 
 import (
 	"bytes"
+	"context"
+	"errors"
 	"fmt"
+	"path/filepath"
 	"strings"
+	"sync"
+	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/bundle"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/config"
-	"github.com/Jeffail/benthos/v3/lib/manager"
+	"github.com/Jeffail/benthos/v3/lib/stream"
 	"github.com/Jeffail/gabs/v2"
+	"github.com/fsnotify/fsnotify"
 	"gopkg.in/yaml.v3"
 )
+
+type configFileInfo struct {
+	updatedAt time.Time
+}
+
+type streamFileInfo struct {
+	configFileInfo
+
+	id string
+}
 
 // Reader provides utilities for parsing a Benthos config as a main file with
 // a collection of resource files, and options such as overrides.
 type Reader struct {
+	// The suffix given to unit test definition files, this is used in order to
+	// exclude unit tests from being run in streams mode with arbitrary
+	// directory walking.
+	testSuffix string
+
 	mainPath      string
 	resourcePaths []string
+	streamsPaths  []string
 	overrides     []string
+
+	// Controls whether the main config should include input, output, etc.
+	streamsMode bool
+
+	// Tracks the details of the config file when we last read it.
+	configFileInfo configFileInfo
+
+	// Tracks the details of stream config files when we last read them.
+	streamFileInfo map[string]streamFileInfo
+
+	// Tracks the details of resource config files when we last read them,
+	// including information such as the specific resources that were created
+	// from it.
+	resourceFileInfo    map[string]resourceFileInfo
+	resourceFileInfoMut sync.Mutex
+
+	mainUpdateFn   MainUpdateFunc
+	streamUpdateFn StreamUpdateFunc
+	watcher        *fsnotify.Watcher
 }
 
 // NewReader creates a new config reader.
 func NewReader(mainPath string, resourcePaths []string, opts ...OptFunc) *Reader {
 	r := &Reader{
-		mainPath:      mainPath,
-		resourcePaths: resourcePaths,
+		testSuffix:       "_benthos_test",
+		mainPath:         mainPath,
+		resourcePaths:    resourcePaths,
+		streamFileInfo:   map[string]streamFileInfo{},
+		resourceFileInfo: map[string]resourceFileInfo{},
 	}
 	for _, opt := range opts {
 		opt(r)
@@ -37,12 +82,192 @@ func NewReader(mainPath string, resourcePaths []string, opts ...OptFunc) *Reader
 // OptFunc is an opt function that changes the behaviour of a config reader.
 type OptFunc func(*Reader)
 
+// OptTestSuffix configures the suffix given to unit test definition files, this
+// is used in order to exclude unit tests from being run in streams mode with
+// arbitrary directory walking.
+func OptTestSuffix(suffix string) OptFunc {
+	return func(r *Reader) {
+		r.testSuffix = suffix
+	}
+}
+
 // OptAddOverrides adds one or more override expressions to the config reader,
 // each of the form `path=value`.
 func OptAddOverrides(overrides ...string) OptFunc {
 	return func(r *Reader) {
 		r.overrides = append(r.overrides, overrides...)
 	}
+}
+
+// OptSetStreamPaths marks this config reader as operating in streams mode, and
+// adds a list of paths to obtain individual stream configs from.
+func OptSetStreamPaths(streamsPaths ...string) OptFunc {
+	return func(r *Reader) {
+		r.streamsPaths = streamsPaths
+		r.streamsMode = true
+	}
+}
+
+//------------------------------------------------------------------------------
+
+// Read a Benthos config from the files and options specified.
+func (r *Reader) Read(conf *config.Type) (lints []string, err error) {
+	if lints, err = r.readMain(conf); err != nil {
+		return
+	}
+	var rLints []string
+	if rLints, err = r.readResources(&conf.ResourceConfig); err != nil {
+		return
+	}
+	lints = append(lints, rLints...)
+	return
+}
+
+// ReadStreams attempts to read Benthos stream configs from one or more paths.
+// Stream configs are extracted and added to a provided map, where the id is
+// derived from the path of the stream config file.
+func (r *Reader) ReadStreams(confs map[string]stream.Config) (lints []string, err error) {
+	return r.readStreamFiles(confs)
+}
+
+// MainUpdateFunc is a closure function called whenever a main config has been
+// updated. A boolean should be returned indicating whether the stream was
+// successfully updated, if false then the attempt will be made again after a
+// grace period.
+type MainUpdateFunc func(conf stream.Config) bool
+
+// SubscribeConfigChanges registers a closure function to be called whenever the
+// main configuration file is updated.
+//
+// The provided closure should return true if the stream was successfully
+// replaced.
+func (r *Reader) SubscribeConfigChanges(fn MainUpdateFunc) error {
+	if r.watcher != nil {
+		return errors.New("a file watcher has already been started")
+	}
+
+	r.mainUpdateFn = fn
+	return nil
+}
+
+// StreamUpdateFunc is a closure function called whenever a stream config has
+// been updated. A boolean should be returned indicating whether the stream was
+// successfully updated, if false then the attempt will be made again after a
+// grace period.
+type StreamUpdateFunc func(id string, conf stream.Config) bool
+
+// SubscribeStreamChanges registers a closure to be called whenever the
+// configuration of a stream is updated.
+//
+// The provided closure should return true if the stream was successfully
+// replaced.
+func (r *Reader) SubscribeStreamChanges(fn StreamUpdateFunc) error {
+	if r.watcher != nil {
+		return errors.New("a file watcher has already been started")
+	}
+
+	r.streamUpdateFn = fn
+	return nil
+}
+
+const (
+	changeFlushPeriod = 50 * time.Millisecond
+	changeDelayPeriod = time.Second
+)
+
+// BeginFileWatching creates a goroutine that watches all active configuration
+// files for changes. If a resource is changed then it is swapped out
+// automatically through the provided manager. If a main config or stream config
+// changes then the closures registered with either SubscribeConfigChanges or
+// SubscribeStreamChanges will be called.
+//
+// WARNING: Either SubscribeConfigChanges or SubscribeStreamChanges must be
+// called before this, as otherwise it is unsafe to register them during
+// watching.
+func (r *Reader) BeginFileWatching(mgr bundle.NewManagement, strict bool) error {
+	if r.watcher != nil {
+		return errors.New("a file watcher has already been started")
+	}
+	if r.mainUpdateFn == nil && r.streamUpdateFn == nil {
+		return errors.New("a file watcher cannot be started without a subscription function registered")
+	}
+
+	watcher, err := fsnotify.NewWatcher()
+	if err != nil {
+		return err
+	}
+	r.watcher = watcher
+
+	go func() {
+		ticker := time.NewTicker(changeFlushPeriod)
+		defer ticker.Stop()
+
+		collapsedChanges := map[string]time.Time{}
+		for {
+			select {
+			case event, ok := <-watcher.Events:
+				if !ok {
+					return
+				}
+				if event.Op&fsnotify.Write != fsnotify.Write {
+					continue
+				}
+				collapsedChanges[filepath.Clean(event.Name)] = time.Now()
+			case <-ticker.C:
+				for nameClean, changed := range collapsedChanges {
+					if time.Since(changed) < changeDelayPeriod {
+						continue
+					}
+					var succeeded bool
+					if nameClean == filepath.Clean(r.mainPath) {
+						succeeded = r.reactMainUpdate(mgr, strict)
+					} else if _, exists := r.streamFileInfo[nameClean]; exists {
+						succeeded = r.reactStreamUpdate(mgr, strict, nameClean)
+					} else {
+						succeeded = r.reactResourceUpdate(mgr, strict, nameClean)
+					}
+					if succeeded {
+						delete(collapsedChanges, nameClean)
+					} else {
+						collapsedChanges[nameClean] = time.Now()
+					}
+				}
+			case err, ok := <-watcher.Errors:
+				if !ok {
+					return
+				}
+				mgr.Logger().Errorf("Config watcher error: %v", err)
+			}
+		}
+	}()
+
+	if !r.streamsMode && r.mainPath != "" {
+		if err := watcher.Add(r.mainPath); err != nil {
+			_ = watcher.Close()
+			return err
+		}
+	}
+	for _, p := range r.streamsPaths {
+		if err := watcher.Add(p); err != nil {
+			_ = watcher.Close()
+			return err
+		}
+	}
+	for _, p := range r.resourcePaths {
+		if err := watcher.Add(p); err != nil {
+			_ = watcher.Close()
+			return err
+		}
+	}
+	return nil
+}
+
+// Close the reader, when this method exits all reloading will be stopped.
+func (r *Reader) Close(ctx context.Context) error {
+	if r.watcher != nil {
+		return r.watcher.Close()
+	}
+	return nil
 }
 
 //------------------------------------------------------------------------------
@@ -93,7 +318,19 @@ func (r *Reader) readMain(conf *config.Type) (lints []string, err error) {
 		}
 	}
 
+	// This is an unlikely race condition as the file could've been updated
+	// exactly when we were reading/linting. However, we'd need to fork
+	// ReadWithJSONPointersLinted in order to pull the file info out, and since
+	// it's going to be removed in V4 I'm just going with the simpler option for
+	// now (ignoring the issue).
+	r.configFileInfo.updatedAt = time.Now()
+
 	confSpec := config.Spec()
+	if r.streamsMode {
+		// Spec is limited to just non-stream fields when in streams mode (no
+		// input, output, etc)
+		confSpec = config.SpecWithoutStream()
+	}
 	if err = applyOverrides(confSpec, &rawNode, r.overrides...); err != nil {
 		return
 	}
@@ -112,57 +349,32 @@ func (r *Reader) readMain(conf *config.Type) (lints []string, err error) {
 	return
 }
 
-func (r *Reader) readResources(conf *config.Type) (lints []string, err error) {
-	for _, path := range r.resourcePaths {
-		rconf := manager.NewResourceConfig()
-		var rLints []string
-		if rLints, err = readResource(path, &rconf); err != nil {
-			return
-		}
-		lints = append(lints, rLints...)
-		if err = conf.ResourceConfig.AddFrom(&rconf); err != nil {
-			err = fmt.Errorf("%v: %w", path, err)
-			return
-		}
-	}
-	return
-}
-
-func readResource(path string, conf *manager.ResourceConfig) (lints []string, err error) {
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("%v: %w", path, err)
-		}
-	}()
-
-	var confBytes []byte
-	if confBytes, lints, err = config.ReadWithJSONPointersLinted(path, true); err != nil {
-		return
+func (r *Reader) reactMainUpdate(mgr bundle.NewManagement, strict bool) bool {
+	if r.mainUpdateFn == nil {
+		return true
 	}
 
-	var rawNode yaml.Node
-	if err = yaml.Unmarshal(confBytes, &rawNode); err != nil {
-		return
-	}
-	if !bytes.HasPrefix(confBytes, []byte("# BENTHOS LINT DISABLE")) {
-		for _, lint := range manager.Spec().LintYAML(docs.NewLintContext(), &rawNode) {
-			lints = append(lints, fmt.Sprintf("resource file %v: line %v: %v", path, lint.Line, lint.What))
-		}
+	mgr.Logger().Infoln("Stream config updated, attempting to update stream.")
+
+	conf := config.New()
+	lints, err := r.readMain(&conf)
+	if err != nil {
+		mgr.Logger().Errorf("Failed to read updated config: %v", err)
+
+		// Rejecting due to invalid file means we do not want to try again.
+		return true
 	}
 
-	err = rawNode.Decode(conf)
-	return
-}
+	lintlog := mgr.Logger().NewModule(".linter")
+	for _, lint := range lints {
+		lintlog.Infoln(lint)
+	}
+	if strict && len(lints) > 0 {
+		mgr.Logger().Errorln("Rejecting updated stream config due to linter errors, to allow linting errors run Benthos with --chilled")
 
-// Read a Benthos config from the files and options specified.
-func (r *Reader) Read(conf *config.Type) (lints []string, err error) {
-	if lints, err = r.readMain(conf); err != nil {
-		return
+		// Rejecting from linters means we do not want to try again.
+		return true
 	}
-	var rLints []string
-	if rLints, err = r.readResources(conf); err != nil {
-		return
-	}
-	lints = append(lints, rLints...)
-	return
+
+	return r.mainUpdateFn(conf.Config)
 }