+++ lib/stream/manager/from_path.go
@@ -6,53 +6,39 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/Jeffail/benthos/v3/lib/config"
+	"github.com/Jeffail/benthos/v3/internal/config"
 	"github.com/Jeffail/benthos/v3/lib/stream"
 )
 
-//------------------------------------------------------------------------------
-
 func loadFile(dir, path, testSuffix string, confs map[string]stream.Config) ([]string, error) {
-	var id string
-	if len(dir) > 0 {
-		var err error
-		if id, err = filepath.Rel(dir, path); err != nil {
-			return nil, err
-		}
-	} else {
-		id = filepath.Base(path)
+	id, err := config.InferStreamID(dir, path)
+	if err != nil {
+		return nil, err
 	}
-	id = strings.Trim(id, string(filepath.Separator))
-	id = strings.TrimSuffix(id, ".yaml")
-	id = strings.TrimSuffix(id, ".yml")
 
 	// Do not run unit test files
 	if len(testSuffix) > 0 && strings.HasSuffix(id, testSuffix) {
 		return nil, nil
 	}
 
-	id = strings.ReplaceAll(id, string(filepath.Separator), "_")
-
 	if _, exists := confs[id]; exists {
 		return nil, fmt.Errorf("stream id (%v) collision from file: %v", id, path)
 	}
 
-	conf := config.New()
-	lints, err := config.Read(path, true, &conf)
+	conf, lints, err := config.ReadStreamFile(path)
 	if err != nil {
 		return nil, err
 	}
-	for i := range lints {
-		lints[i] = path + ": " + lints[i]
-	}
 
-	confs[id] = conf.Config
+	confs[id] = conf
 	return lints, nil
 }
 
 // LoadStreamConfigsFromPath reads a map of stream ids to configurations
 // by either walking a directory of .json and .yaml files or by reading a file
 // directly. Returns linting errors prefixed with their path.
+//
+// Deprecated: The streams builder is using ./internal/config now.
 func LoadStreamConfigsFromPath(target, testSuffix string, streamMap map[string]stream.Config) ([]string, error) {
 	pathLints := []string{}
 	target = filepath.Clean(target)
@@ -87,5 +73,3 @@ func LoadStreamConfigsFromPath(target, testSuffix string, streamMap map[string]s
 
 	return pathLints, err
 }
-
-//------------------------------------------------------------------------------